% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec.R
\name{dec}
\alias{dec}
\title{Unpack shipment and decrypt content}
\usage{
dec(tarfile, keyfile = "~/.ssh/id_rsa", target_dir = ".")
}
\arguments{
\item{tarfile}{Character string providing full path to the gzip-compressd
tarball holding the shipment payload, including encrypted files}

\item{keyfile}{Character string providing the full path to the private RSA
key to be used for decryption of the encrypted key that is part of the
shipment. Default value is set to \code{~/.ssh/id_rsa} which is the usual
path for unix-type operating systems.}

\item{target_dir}{Character string providing the full path to where the
decrypted file is to be written. Defaults to the current directory
\code{"."}, \emph{e.g.} where this function is being called from.}
}
\value{
Invisibly the character string file path of the encrypted file.
}
\description{
This function tries to reverse the process of \link{enc_file} and
hence depend on the conventions used there.
}
\details{
Some of the functions used here might be vulnerable to differences between
systems running R. Possible caveats may be the availability of the
(un)tar-function and how binary streams/files are treated.
}
\seealso{
\link{enc_file}
}
