% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_api}
\alias{github_api}
\title{Make calls to the github API}
\usage{
github_api(path, proxy_url = NULL)
}
\arguments{
\item{path}{Character string with path to the API resource}

\item{proxy_url}{Character string defining a network proxy in the form
host:port. Default is NULL in which case the API call will not use a proxy}
}
\value{
A list of class github_api containing the parsed content, API
resource path and the response object
}
\description{
Make calls to the github API
}
\examples{
github_api("repos/Rapporteket/sship/branches")
}
