% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec.R
\name{sym_dec}
\alias{sym_dec}
\title{Standard sship symmetric decryption}
\usage{
sym_dec(data, key, iv = attr(data, "iv"))
}
\arguments{
\item{data}{raw vector or path to file with data to encrypt or decrypt}

\item{key}{raw vector of length 16, 24 or 32, e.g. the hash of a shared
secret}

\item{iv}{raw vector of length 16 (aes block size) or NULL. The
initialization vector is not secret but should be random}
}
\value{
A raw vector of decrypted \code{data}.
}
\description{
Standard sship symmetric decryption
}
\keyword{internal}
