% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enc.R
\name{sym_enc}
\alias{sym_enc}
\title{Standard sship symmetric encryption}
\usage{
sym_enc(data, key, iv = openssl::rand_bytes(16))
}
\arguments{
\item{data}{raw vector or path to file with data to encrypt or decrypt}

\item{key}{raw vector of length 16, 24 or 32, e.g. the hash of a shared
secret}

\item{iv}{raw vector of length 16 (aes block size) or NULL. The
initialization vector is not secret but should be random}
}
\value{
A raw vector of encrypted \code{data}.
}
\description{
Standard sship symmetric encryption
}
\keyword{internal}
